reorganization of team headers
